+++ internal/infra/table/extractor_db2.go
@@ -72,3 +72,7 @@ func (d Db2Dialect) SQL(schema string) string {
 
 	return SQL
 }
+
+func (d Db2Dialect) GetExportType(dbtype string) (string, bool) {
+	return "", false
+}