+++ internal/infra/table/extractor_db2_dummy.go
@@ -47,3 +47,7 @@ type Db2Dialect struct {
 func (d Db2Dialect) SQL(schema string) string {
 	panic(fmt.Errorf("Not implemented"))
 }
+
+func (d Db2Dialect) GetExportType(dbtype string) (string, bool) {
+	return "", false
+}